requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

fake-useragent
beautifulsoup4
azure.identity
azure.keyvault
azure-functions
pandas
numpy
requests
regex
datetime